Fix all the bugs

+++ DatabaseReader.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -30,18 +31,18 @@ namespace MultimediaRetrieval
         {
             DatabaseReader total = new DatabaseReader();
 
-            string[] files = Directory.GetFiles(dirpath, "*.mr", SearchOption.TopDirectoryOnly);
+            string[] files = Directory.GetFiles(dirpath, "*.mr");
             if (files.Length > 0)
             {
                 throw new System.ArgumentException("This folder contains already a statistic file");
             }
 
-            files = Directory.GetFiles(dirpath, "*.cla", SearchOption.TopDirectoryOnly);
+            files = Directory.GetFiles(dirpath, "*.cla");
             if (files.Length > 0)
             {
                 foreach (string filepath in files)
                 {
-                    ReadClassificationPrinceton(total, dirpath, dirpath + filepath, outpath);
+                    ReadClassificationPrinceton(total, dirpath, filepath, outpath);
                 }
                 return total;
             }
@@ -78,13 +79,15 @@ namespace MultimediaRetrieval
                     if (line == string.Empty)
                         continue;
 
+                    if (line.StartsWith("PSB", System.StringComparison.InvariantCulture))
+                        continue;
+
                     string[] split = line.Split(new char[] { ' ' }, System.StringSplitOptions.RemoveEmptyEntries);
 
-                    // If it is any other line, such as header, ignore it.
                     if (split.Length != 3)
                         continue;
 
-                    string cls = split[0];
+                    string cls = split[0].ToLower();
                     int amount = int.Parse(split[2]);
 
                     for(int i = 0; i < amount; i++)
@@ -94,6 +97,7 @@ namespace MultimediaRetrieval
                     }
                 }
             }
+
 
             foreach (string filename in ListMeshes(dirpath))
             {
@@ -110,15 +114,17 @@ namespace MultimediaRetrieval
 
         public static IEnumerable<string> ListMeshes(string dirpath)
         {
-            foreach (string filename in Directory.EnumerateFiles(dirpath))
+            foreach (string filename in Directory.EnumerateFiles(dirpath, "*", SearchOption.AllDirectories))
             {
-                string fullname = dirpath + filename;
-                string extension = Path.GetExtension(fullname);
+                string extension = Path.GetExtension(filename);
 
-                if (UNSUPPORTED_EXTENSIONS.Any((s) => s == extension))
+                if (extension.EndsWith("~", StringComparison.InvariantCulture))
                     continue;
 
-                yield return fullname;                
+                if (UNSUPPORTED_EXTENSIONS.Any((s) => string.Equals(s, extension)))
+                    continue;
+
+                yield return filename;                
             }
         }
 
@@ -130,7 +136,8 @@ namespace MultimediaRetrieval
         private static void MoveFile(string filename, string outpath, uint id)
         {
             string newfile = outpath + Path.DirectorySeparatorChar + id + Path.GetExtension(filename);
-            File.Copy(filename, newfile, false);
+            if (!File.Exists(newfile))
+                File.Copy(filename, newfile, false);
         }
     }
 }
